cron

+++ entrypoint.sh
@@ -2,26 +2,36 @@
 
 set -e
 
-echo "Entrypoint script started as user: $(whoami)"
+echo "Entrypoint script started..."
 
-echo "Taking ownership of volume directories..."
-chown -R appuser:appuser /app/media
-chown -R appuser:appuser /app/staticfiles
+if [ "$1" = "/app/cron.sh" ]; then
+    echo "Running in Cron mode (staying as root to configure daemon)..."
+else
+    echo "Entrypoint script started as user: $(whoami)"
+    echo "Taking ownership of volume directories..."
+    chown -R appuser:appuser /app/media
+    chown -R appuser:appuser /app/staticfiles
+fi
 
 echo "Setting mask..."
 umask 0002
 
-echo "Applying database migrations..."
-gosu appuser python manage.py migrate --noinput
+if [ "$1" != "/app/cron.sh" ]; then
+    echo "Applying database migrations..."
+    gosu appuser python manage.py migrate --noinput
 
-echo "Wiping old static files from volume..."
-gosu appuser find /app/staticfiles/ -mindepth 1 -delete
+    echo "Wiping old static files from volume..."
+    gosu appuser find /app/staticfiles/ -mindepth 1 -delete
 
-echo "Compiling tailwind///"
-python manage.py tailwind build
+    echo "Compiling tailwind..."
+    python manage.py tailwind build
 
-echo "Collecting static files for production..."
-gosu appuser python manage.py collectstatic --noinput --clear
-
-echo "Starting application server as user: $(whoami)..."
-exec gosu appuser "$@"
+    echo "Collecting static files for production..."
+    gosu appuser python manage.py collectstatic --noinput --clear
+    
+    echo "Starting application server as user: appuser..."
+    exec gosu appuser "$@"
+else
+    echo "Starting Cron..."
+    exec "$@"
+fi
